requirements.txt
fake-bpy-module-2.80
PySide2
#pypiwin32